+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @varveai/adit-cli
 
+## 0.3.2
+
+### Patch Changes
+
+- 部分重构
+- Updated dependencies
+  - @varveai/adit-engine@0.3.2
+  - @varveai/adit-cloud@0.3.2
+  - @varveai/adit-hooks@0.3.2
+  - @varveai/adit-plans@0.3.2
+  - @varveai/adit-core@0.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- Daily update
+- Updated dependencies
+  - @varveai/adit-engine@0.3.1
+  - @varveai/adit-cloud@0.3.1
+  - @varveai/adit-hooks@0.3.1
+  - @varveai/adit-plans@0.3.1
+  - @varveai/adit-core@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes